+++ packages/react-components/react-migration-v0-v9/src/components/Slider/Slider.mixins.ts
@@ -1,0 +1,7 @@
+import { GriffelStyle } from '@fluentui/react-components';
+
+const fluid = (): GriffelStyle => ({ width: '100%' });
+
+export const slider = {
+  fluid,
+};