+++ packages/react-components/react-utilities/src/hooks/useId.ts
@@ -34,10 +34,15 @@ export function useId(prefix: string = 'fui-', providedId?: string): string {
 
   // Checking if useId is available on React, if it is, we use it to generate the id. String concatenation is used to
   // prevent bundlers from complaining with older versions of React.
-  const _useId: () => string | undefined = (React as never)['use' + 'Id'];
+  const _useId = (React as never)['use' + 'Id'] as (() => string) | undefined;
 
   if (_useId) {
-    return providedId || `${idPrefix}${prefix}${_useId()}`;
+    const generatedId = _useId();
+
+    // eslint-disable-next-line react-hooks/rules-of-hooks
+    const escapedId = React.useMemo(() => generatedId.replace(/:/g, ''), [generatedId]);
+
+    return providedId || `${idPrefix}${prefix}${escapedId}`;
   }
 
   // Hooks appear to be running conditionally, but they will always run in the same order since it's based on